fix: auto extract images
